web-scraper/pyproject.toml
[build-system]
requires = ["poetry-core>=2.0"]
build-backend = "poetry.core.masonry.api"

[project]
name = "web_scraper"
readme = "README.md"
requires-python = "~3.12"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.116.1",
    "curl-cffi (>=0.13.0,<0.14.0)",
    "newspaper4k (>=0.9.3.1,<0.10.0.0)",
    "lxml[html-clean] (>=6.0.0,<7.0.0)",
]


[tool.poetry]
packages = [
    { include = "web_scraper", from = "." },
]
version = "0.0.0"


[tool.poetry.group.dev.dependencies]
lxml-stubs = "^0.5.1"

[tool.poetry-monoranger-plugin]
enabled = true
monorepo-root = "../"
version-pinning-rule = '=='  